requirements.txt
asgiref==3.4.1
Django==2.1.15
django-cors-headers==2.4.0
django-pyodbc-azure==2.1.0.0
django-pyodbc-azure-2019==2.1.0.0
djangorestframework==3.8.2
pyodbc==4.0.38
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.1.1
